[frontend] widget perspective type + render icon in utils
