+++ opencti-platform/opencti-front/lang/front/en.json
@@ -2863,7 +2863,6 @@
   "This user has no Max Confidence Level and does not inherit one from groups. Add a group to this user to resolve the issue.": "This user has no Max Confidence Level and does not inherit one from groups. Add a group to this user to resolve the issue.",
   "This value cannot be null": "This value cannot be null",
   "This value is overridden for the following entity types": "This value is overridden for the following entity types",
-  "This widget type is not implemented": "This widget type is not implemented",
   "Threat actor": "Threat actor",
   "Threat actor types": "Threat actor types",
   "Threat actors": "Threat actors",
