+++ opencti-platform/opencti-front/src/public/components/PublicDashboard.tsx
@@ -1,9 +1,10 @@
 import React, { useEffect, useMemo } from 'react';
 import { graphql, PreloadedQuery, usePreloadedQuery } from 'react-relay';
-import { useParams, useNavigate } from 'react-router-dom';
+import { useNavigate, useParams } from 'react-router-dom';
 import RGL, { WidthProvider } from 'react-grid-layout';
 import { ErrorBoundary } from '@components/Error';
 import Paper from '@mui/material/Paper';
+import { WidgetPerspective } from '@components/widgets/widgetUtils';
 import Loader, { LoaderVariant } from '../../components/Loader';
 import { PublicDashboardQuery } from './__generated__/PublicDashboardQuery.graphql';
 import useQueryLoading from '../../utils/hooks/useQueryLoading';
@@ -95,9 +96,9 @@ const PublicDashboardComponent = ({
             }}
           >
             <ErrorBoundary>
-              {widget.perspective === 'entities' && entityWidget(widget)}
-              {widget.perspective === 'relationships' && relationshipWidget(widget)}
-              {widget.perspective === 'audits' && auditWidget(widget)}
+              {widget.perspective === WidgetPerspective.entities && entityWidget(widget)}
+              {widget.perspective === WidgetPerspective.relationships && relationshipWidget(widget)}
+              {widget.perspective === WidgetPerspective.audits && auditWidget(widget)}
               {widget.perspective === null && rawWidget(widget)}
             </ErrorBoundary>
           </Paper>
